Changed maria to write to /scratch instead of /var/*
This will force singularity to be launched with something like
`singularity -B $SCRATCH:/scratch container.img` but should get
around needing the container to write to itself and change its size
thus making it larger than it needs to be. Will need to determine
if mutable container is better for long term storage or not.